Add stub for mouse_over for buttons
The feature is not yet completely implemented, but this commit should make it possible to listen to components as soon as I find out how UIObject is used to load the respective component. The UIObject class now has a MouseListener, because it should have one. Note: some useless updates were also implemented in this commit..