Updated the sprite and added balance changes
The charge time was increased from 5 seconds to 10 seconds (and it is now clearly visible when it has finished charging), but it will not self-detonate. Its damage has also been reduced. It will now do short bursts of 100 damage per zombie in its lane, giving it an average dps of 10 per zombie, which is half of Maxim's EggShooter).

I also renamed the texture files.